+++ src/Sentry/TransactionContext.cs
@@ -3,10 +3,13 @@ namespace Sentry
     /// <summary>
     /// Transaction metadata used for sampling.
     /// </summary>
-    public class TransactionContext : SpanContext, ITransactionContext
+    public class TransactionContext : SpanContext, ITransactionContext, IHasTransactionNameSource
     {
         /// <inheritdoc />
         public string Name { get; }
+
+        /// <inheritdoc />
+        public TransactionNameSource NameSource { get; }
 
         /// <summary>
         /// Whether the parent transaction of this transaction has been sampled.
@@ -25,11 +28,33 @@ namespace Sentry
             string? description,
             SpanStatus? status,
             bool? isSampled,
+            bool? isParentSampled,
+            TransactionNameSource nameSource)
+            : base(spanId, parentSpanId, traceId, operation, description, status, isSampled)
+        {
+            Name = name;
+            IsParentSampled = isParentSampled;
+            NameSource = nameSource;
+        }
+
+        /// <summary>
+        /// Initializes an instance of <see cref="TransactionContext"/>.
+        /// </summary>
+        public TransactionContext(
+            SpanId spanId,
+            SpanId? parentSpanId,
+            SentryId traceId,
+            string name,
+            string operation,
+            string? description,
+            SpanStatus? status,
+            bool? isSampled,
             bool? isParentSampled)
             : base(spanId, parentSpanId, traceId, operation, description, status, isSampled)
         {
             Name = name;
             IsParentSampled = isParentSampled;
+            NameSource = TransactionNameSource.Custom;
         }
 
         /// <summary>
@@ -63,7 +88,19 @@ namespace Sentry
             string name,
             string operation,
             SentryTraceHeader traceHeader)
-            : this(traceHeader.SpanId, traceHeader.TraceId, name, operation, traceHeader.IsSampled)
+            : this(SpanId.Create(), traceHeader.SpanId, traceHeader.TraceId, name, operation, "", null, traceHeader.IsSampled, traceHeader.IsSampled)
+        {
+        }
+
+        /// <summary>
+        /// Initializes an instance of <see cref="TransactionContext"/>.
+        /// </summary>
+        public TransactionContext(
+            string name,
+            string operation,
+            SentryTraceHeader traceHeader,
+            TransactionNameSource nameSource)
+            : this(SpanId.Create(), traceHeader.SpanId, traceHeader.TraceId, name, operation, "", null, traceHeader.IsSampled, traceHeader.IsSampled, nameSource)
         {
         }
 
@@ -71,7 +108,15 @@ namespace Sentry
         /// Initializes an instance of <see cref="TransactionContext"/>.
         /// </summary>
         public TransactionContext(string name, string operation)
-            : this(name, operation, (bool?)null)
+            : this(SpanId.Create(), null, SentryId.Create(), name, operation, "", null, null, null)
+        {
+        }
+
+        /// <summary>
+        /// Initializes an instance of <see cref="TransactionContext"/>.
+        /// </summary>
+        public TransactionContext(string name, string operation, TransactionNameSource nameSource)
+            : this(SpanId.Create(), null, SentryId.Create(), name, operation, "", null, null, null, nameSource)
         {
         }
     }