+++ test/Sentry.AspNet.Tests/ApiApprovalTests.cs
@@ -1,5 +1,6 @@
-using Sentry.AspNet;
 using Sentry.Tests;
+
+namespace Sentry.AspNet.Tests;
 
 [UsesVerify]
 public class ApiApprovalTests